backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
ebooklib>=0.18
beautifulsoup4>=4.12.0
google-genai>=0.8.0
lxml>=5.0.0
pydantic>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[sha256_crypt]>=1.7.4
